Add support for #pragma pack
This uses the existing logic for handeling packed structs to also pack
structs when `pragma pack` is used.

Add support for #pragma pack

+++ src/ast.zig
@@ -645,7 +645,7 @@ pub const Payload = struct {
         pub const Field = struct {
             name: []const u8,
             type: Node,
-            alignment: ?c_uint,
+            alignment: ?usize,
             default_value: ?Node,
         };
     };
